+++ src/solvers/test/mechanics/mechanicsVerification.cpp
@@ -160,7 +160,8 @@ computeExactSolution( AMP::Mesh::Mesh::shared_ptr meshAdapter,
             manufacturedSolution->getExactSolutions( coord[0], coord[1], coord[2] );
         // Distribute values in the vector object
         for ( unsigned int xyz = 0; xyz < 3; ++xyz ) {
-	  exactSolutionsVec->setLocalValuesByGlobalID( 1, &globalIDs[xyz], &displacementXYZ[xyz] );
+            exactSolutionsVec->setLocalValuesByGlobalID(
+                1, &globalIDs[xyz], &displacementXYZ[xyz] );
         } // end loop over the coordinates
     }     // end soop over all nodes
     if ( verbose ) {
@@ -413,7 +414,8 @@ static void linearElasticTest( AMP::UnitTest *ut, std::string exeName, int examp
         ut->passes(exeName);
      } */
 
-    double epsilon = 1.0e-13 * ( ( ( bvpOperator->getMatrix() )->extractDiagonal() )->L1Norm() );
+    double epsilon = 1.0e-13 * static_cast<double>(
+                                   ( ( bvpOperator->getMatrix() )->extractDiagonal() )->L1Norm() );
     AMP::pout << "epsilon = " << epsilon << std::endl;
 
     AMP::pout << "------------------------------------------------\n"
