+++ src/solvers/test/petsc/testPetscSNESSolver.cpp
@@ -69,7 +69,7 @@ static void myTest( AMP::UnitTest *ut, const std::string &exeName )
     ut->passes( "PetscSNESSolver solve called with simple vector" );
     auto x = u->cloneVector();
     x->subtract( *u, *f );
-    double error = x->L2Norm() / f->L2Norm();
+    double error = static_cast<double>( x->L2Norm() ) / static_cast<double>( f->L2Norm() );
     if ( fabs( error ) < 1e-8 )
         ut->passes( "Solve with simple vector passed" );
     else
